Support for dtnd Time-Synchronization enable
When using the dtnd daemon with time synchronization enabled the slave
will send bundles with the timestamp set to 0 and add to the front a
AGE_BLOCK.

When creating the bundle we must let the timestamp given, even if it is
zero (which could mean that it was send by a node in slave mode).
When reading the bundle info we must not load the AGE_BLOCK since it will raise an API_STATUS_NOT_ACCEPTABLE.

We thus need to parse each block and only load those with a BLOCK_TYPE=1
(payload block type).